# frontend/app/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
    "/api/v1/contracts/{id}/mute" : {
      "put" : {
        "tags" : [ "contract-controller" ],
        "operationId" : "mute",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MuteRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "contract-controller" ],
        "operationId" : "unmute",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/contracts" : {
      "get" : {
        "tags" : [ "contract-controller" ],
        "operationId" : "list_2",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "ALL",
            "enum" : [ "ALL", "HEARTBEAT_STOPPED", "HEARTBEAT_NEVER", "FULL_SYNC_OVERDUE", "FULL_SYNC_NEVER", "MUTED" ]
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "HEALTH",
            "enum" : [ "HEALTH", "ORG", "USERNAME", "HEARTBEAT", "FULL_SYNC", "CAPABILITIES" ]
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "ASC",
            "enum" : [ "ASC", "DESC" ]
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{id}" : {
      "get" : {
        "tags" : [ "contract-controller" ],
        "operationId" : "get_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "MuteRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : [ "string", "null" ]
          },
          "mutedUntil" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        }
      },
      "CapabilityView" : {
        "type" : "object",
        "properties" : {
          "domainName" : {
            "type" : "string"
          },
          "packageName" : {
            "type" : "string"
          },
          "resourceName" : {
            "type" : "string"
          },
          "fullSyncIntervalDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deltaSyncInterval" : {
            "type" : [ "string", "null" ]
          },
          "lastFullSyncAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "fullSync" : {
            "type" : "string",
            "enum" : [ "OK", "WAITING", "OVERDUE", "NEVER" ]
          }
        }
      },
      "ContractDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orgId" : {
            "type" : "string"
          },
          "mainOrgId" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "adapterId" : {
            "type" : "string"
          },
          "heartbeatIntervalMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "firstRegisteredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "registeredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastHeartbeatAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "heartbeat" : {
            "type" : "string",
            "enum" : [ "OK", "STOPPED", "NEVER" ]
          },
          "mute" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/MuteView"
            }, {
              "type" : "null"
            } ]
          },
          "capabilities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CapabilityView"
            }
          }
        }
      },
      "MuteView" : {
        "type" : "object",
        "properties" : {
          "mutedBy" : {
            "type" : "string"
          },
          "reason" : {
            "type" : [ "string", "null" ]
          },
          "mutedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "mutedUntil" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        }
      },
...
      "ContractPage" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ContractSummary"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "counts" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }
      },
      "ContractSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "orgId" : {
            "type" : "string"
          },
          "mainOrgId" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "heartbeat" : {
            "type" : "string",
            "enum" : [ "OK", "STOPPED", "NEVER" ]
          },
          "lastHeartbeatAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "fullSyncOverdue" : {
            "type" : "integer",
            "format" : "int32"
          },
          "fullSyncNever" : {
            "type" : "integer",
            "format" : "int32"
          },
          "capabilities" : {
            "type" : "integer",
            "format" : "int32"
          },
          "muted" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}